Fix failing UI test, add unit test and refactor
